chore: 별점타입 정수로 변경

+++ src/main/java/org/chunsik/pq/feedback/dto/FeedbackDTO.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 @Getter
 public class FeedbackDTO {
     private Integer userId;
-    private Float starRate;
+    private Integer starRate;
     private String comment;
     private Integer ease;
     private Integer design;
